+++ sdk/storage/storage-blob/src/BlobServiceClient.ts
@@ -22,8 +22,6 @@ import {
   ContainerItem,
   ListContainersIncludeType,
   UserDelegationKeyModel,
-  ServiceFindBlobsByTagsSegmentResponse,
-  FilterBlobItem,
   ContainerUndeleteResponse
 } from "./generatedModels";
 import { Container, Service } from "./generated/src/operations";
@@ -32,7 +30,8 @@ import { ContainerClient, ContainerCreateOptions, ContainerDeleteMethodOptions }
 import {
   appendToURLPath,
   appendToURLQuery,
-  extractConnectionStringParts
+  extractConnectionStringParts,
+  toTags
 } from "./utils/utils.common";
 import { StorageSharedKeyCredential } from "./credentials/StorageSharedKeyCredential";
 import { AnonymousCredential } from "./credentials/AnonymousCredential";
@@ -42,6 +41,8 @@ import { truncatedISO8061Date } from "./utils/utils.common";
 import { createSpan } from "./utils/tracing";
 import { BlobBatchClient } from "./BlobBatchClient";
 import { CommonOptions, StorageClient } from "./StorageClient";
+import { Tags } from "./models";
+import { ServiceFilterBlobsResponse } from "./generated/src/models";
 import { AccountSASPermissions } from "./sas/AccountSASPermissions";
 import { SASProtocol } from "./sas/SASQueryParameters";
 import { SasIPRange } from "./sas/SasIPRange";
@@ -197,6 +198,51 @@ interface ServiceFindBlobsByTagsSegmentOptions extends CommonOptions {
    */
   maxPageSize?: number;
 }
+
+/**
+ * Blob info from a {@link BlobServiceClient.findBlobsByTags}
+ */
+export interface FilterBlobItem {
+  /**
+   * Blob Name.
+   *
+   * @type {string}
+   * @memberof FilterBlobItem
+   */
+  name: string;
+
+  /**
+   * Container Name.
+   *
+   * @type {string}
+   * @memberof FilterBlobItem
+   */
+  containerName: string;
+
+  /**
+   * Blob Tags.
+   *
+   * @type {Tags}
+   * @memberof FilterBlobItem
+   */
+  tags?: Tags;
+
+  /**
+   * Tag value.
+   *
+   * @deprecated The service no longer returns this value. Use {@link tags} to fetch all matching Blob Tags.
+   * @type {string}
+   * @memberof FilterBlobItem
+   */
+  tagValue?: string;
+}
+
+/**
+ * Contains response data for the {@link BlobServiceClient.findBlobsByTags} operation.
+ */
+export type ServiceFindBlobsByTagsSegmentResponse = Omit<ServiceFilterBlobsResponse, "blobs"> & {
+  blobs: FilterBlobItem[];
+};
 
 /**
  * Options to configure the {@link BlobServiceClient.listContainers} operation.
@@ -868,20 +914,32 @@ export class BlobServiceClient extends StorageClient {
     marker?: string,
     options: ServiceFindBlobsByTagsSegmentOptions = {}
   ): Promise<ServiceFindBlobsByTagsSegmentResponse> {
-    // TODO: Rename response.blobs to response.blobItems?
     const { span, spanOptions } = createSpan(
       "BlobServiceClient-findBlobsByTagsSegment",
       options.tracingOptions
     );
 
     try {
-      return await this.serviceContext.filterBlobs({
+      const response = await this.serviceContext.filterBlobs({
         abortSignal: options.abortSignal,
         where: tagFilterSqlExpression,
         marker,
         maxPageSize: options.maxPageSize,
         spanOptions
       });
+
+      const wrappedResponse: ServiceFindBlobsByTagsSegmentResponse = {
+        ...response,
+        _response: response._response, // _response is made non-enumerable
+        blobs: response.blobs.map((blob) => {
+          let tagValue = undefined;
+          if (blob.tags?.blobTagSet.length === 1) {
+            tagValue = blob.tags.blobTagSet[0].value;
+          }
+          return { ...blob, tags: toTags(blob.tags), tagValue };
+        })
+      };
+      return wrappedResponse;
     } catch (e) {
       span.setStatus({
         code: CanonicalCode.UNKNOWN,